% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.r
\name{allocate}
\alias{allocate}
\title{Allocate samples to soil classes}
\usage{
allocate(composition, polygon, n = 15, method = "weighted")
}
\arguments{
\item{composition}{A \code{data.frame} containing the soil class composition of all polygons in a choropleth soil map.}

\item{polygon}{The id of the polygon from which to draw the soil classes.}

\item{n}{The number of samples to draw.}

\item{method}{The method of allocation. Permissible values are \code{"weighted"} (the default), \code{"random-mapunit"} and
\code{"random-all"}. See \emph{Details} for information on how the modes operate.}
}
\description{
\code{allocate} generates a character vector of soil class codes of length \code{n}. The soil classes belong to the soil map polygon
with id \code{polygon} as described in \code{composition}. The soil classes can be sampled by several different methods, as described in
\emph{Details}.
}
\details{
The default allocation method is \code{method = "weighted"}, which weighted-randomly samples the soil classes of a given polygon
using the soil classes' proportions of occurrence as weights. Thus if three soil classes A, B and C are in a soil map unit in
proportions of 70\%, 20\% and 10\% respectively, at each draw, class A would have a 70\% chance of being selected at random.

With \code{method = "random-mapunit"}, classes are drawn at complete random from among those that exist in the given polygon. This
effectively disregards the proportion of the map unit that each class is assumed to occur in. \code{method = "random-all"} works
the same way except the classes are drawn from among all those in the whole soil map, regardless of whether they occur in the given polygon
or not.
}
