% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{confusion_index}
\alias{confusion_index}
\title{Compute confusion index}
\usage{
confusion_index(r, cpus)
}
\arguments{
\item{r}{A RasterStack of probabilities, where each layer corresponds to a
different soil class.}

\item{cpus}{An integer that identifies the number of CPU nodes for parallel
processing.}
}
\value{
A \code{RasterLayer} containing the confusion index data.
}
\description{
This function computes the confusion index. It first sorts the probabilities
for each grid cell in descending order using \code{\link{sort_stack_values}}.
}
