% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dalrymple_covariates}
\alias{dalrymple_covariates}
\title{\emph{scorpan} environmental covariates for a small part of the former
Dalrymple Shire, Queensland, Australia.}
\format{
An object of class \code{RasterStack} of dimension 235 x 234 x 20.
}
\usage{
dalrymple_covariates
}
\description{
\emph{scorpan} environmental covariates for a small part of the former
Dalrymple Shire, Queensland, Australia.
}
\keyword{datasets}
