% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dalrymple_lookup}
\alias{dalrymple_lookup}
\title{Lookup table that relates soil class codes in the soil map unit composition
to integer codes in the soil class realisation rasters.}
\format{
An object of class \code{data.frame} with 71 rows and 2 columns.
}
\usage{
dalrymple_lookup
}
\description{
Lookup table that relates soil class codes in the soil map unit composition
to integer codes in the soil class realisation rasters.
}
\keyword{datasets}
