% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dalrymple_polygons}
\alias{dalrymple_polygons}
\title{A small part of the Land Resources of the Dalrymple Shire soil map from
Queensland, Australia.}
\format{
An object of class \code{PackedSpatVector} of length 1.
}
\usage{
dalrymple_polygons
}
\description{
A small part of the Land Resources of the Dalrymple Shire soil map from
Queensland, Australia.
}
\keyword{datasets}
