% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{.blocks_per_node}
\alias{.blocks_per_node}
\title{Compute raster::clusterR tuning parameter m}
\usage{
.blocks_per_node(rows, cols, cpus, max_cells = 1e+06)
}
\arguments{
\item{rows}{An integer that identifies the number of rows in the target grid.}

\item{cols}{An integer that identifies the number of columns in the target
grid.}

\item{cpus}{An integer that identifies the number of CPU nodes for parallel
processing.}

\item{max_cells}{An integer that identifies the number of cells allowed in
each block.}
}
\value{

}
\description{
This function computes a value for the \code{\link[raster]{clusterR}} tuning
parameter \code{m}.
}
\details{
\code{clusterR} subdivides a large raster into a set of tiles where each tile
has the same number of columns as the original raster, but a smaller number
of rows. The total number of tiles is equal to \code{m} * \code{n}, where
\code{n} is the number of CPU cores (see \code{\link[raster]{beginCluster}})
and \code{m} is the number of tiles per CPU core.

This function calculates the number of tiles per CPU core, \code{m}, given
the number of CPU cores, the dimensions of the overall raster, and the
desired tile size expressed as the number of grid cells per tile (default
value 1,000,000).

If the number of grid cells in the overall raster is smaller than the desired
tile size, the function returns a value of 2, which is the default value of
\code{m} in \code{clusterR}.
}
