% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{.getStratifiedVirtualSamples}
\alias{.getStratifiedVirtualSamples}
\title{Sample the polygons of a SpatialPolygonsDataFrame.}
\usage{
.getStratifiedVirtualSamples(
  covariates,
  polygons,
  composition,
  strata,
  n.realisations = 100,
  rate = 15,
  method.sample = "by_polygon",
  method.allocate = "weighted"
)
}
\description{
Sample the polygons of a SpatialPolygonsDataFrame.
}
