% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{order_cpp}
\alias{order_cpp}
\title{C++ Ordering algorithm}
\usage{
order_cpp(x, decreasing = FALSE)
}
\arguments{
\item{x}{A numeric vector to be ordered}

\item{decreasing}{Set to TRUE or FALSE to order in decreasing order}
}
\description{
C++ Ordering algorithm
}
