% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{shannon_entropy}
\alias{shannon_entropy}
\title{Compute Shannon's entropy}
\usage{
shannon_entropy(r, cpus)
}
\arguments{
\item{r}{A RasterStack of probabilities, where each layer corresponds to a
different soil class.}

\item{cpus}{An integer that identifies the number of CPU nodes for parallel
processing.}
}
\value{
A \code{RasterLayer} containing the Shannon entropy values.
}
\description{
Shannon's entropy is an estimate of the prediction uncertainty.
}
\details{
Shannon's entropy is computed as follows:

\deqn{H=-\sum_{i=1}^{n_\text{orders}}{P_i\log_{n_\text{orders}}P_i}}

where \eqn{P_i} is the probability of occurrence of soil class \eqn{i}. The
logarithm with base \eqn{n_\text{orders}} is used so that the maximum entropy
is 1, which occurs when all soil classes have equal probability of occurrence
(Kempen \emph{et al.}, 2009). The minimum entropy is 0, which occurs when one
soil order has a probability of 1 and all others zero (minimum uncertainty).
}
\references{
Kempen, B., Brus, D.J., Heuvelink, G.B.M., Stoorvogel, J.J.,
  2009. Updating the 1:50,000 Dutch soil map using legacy soil data: a
  multinomial logistic regression approach. Geoderma. 151, 311--326. doi:
  \href{https://doi.org/10.1016/j.geoderma.2009.04.023}{10.1016/j.geoderma.2009.04.023}
}
\seealso{
Other uncertainty functions: 
\code{\link{confusion_index}()}
}
\concept{uncertainty functions}
