% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sort_cpp}
\alias{sort_cpp}
\title{C++ Sorting algorithm}
\usage{
sort_cpp(x, decreasing = FALSE, nalast = FALSE)
}
\arguments{
\item{x}{A numeric vector to be sorted}

\item{decreasing}{set to TRUE or FALSE to sort in decreasing order}

\item{nalast}{Place NA values at the end of the string?}
}
\description{
C++ Sorting algorithm
}
