% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{confusion_index}
\alias{confusion_index}
\title{Compute confusion index}
\usage{
confusion_index(r, cpus)
}
\arguments{
\item{r}{A RasterStack of probabilities, where each layer corresponds to a
different soil class.}

\item{cpus}{An integer that identifies the number of CPU nodes for parallel
processing.}
}
\value{
A \code{RasterLayer} containing the confusion index data.
}
\description{
The confusion index is an estimate of the prediction uncertainty.
}
\details{
The confusion index is computed as follows:

\deqn{C = 1-(P_1-P_2); P_1\geqslant{P_2}}

where \eqn{P_1} is the probability of the most probable class and \eqn{P_2}
is the probability of the second most probable class (after Burrough \emph{et
al.}, 1997; Odgers \emph{et al.}, 2014). The maximum confusion is 1, which
occurs when \eqn{P_1=P_2}. The minimum confusion is 0, which occurs when
\eqn{P_1=1} and \eqn{P_2=0} (and therefore all other \eqn{P_i=0} too).

This function computes the confusion index. It first sorts the probabilities
for each grid cell in descending order using \code{\link{sort_stack_values}}.
}
\references{
Burrough, P.A., van Gaans, P.F.M., Hootsmans, R., 1997.
  Continuous classification in soil survey: spatial correlation, confusion
  and boundaries. Geoderma. 77, 115--135. doi:
  \href{https://doi.org/10.1016/S0016-7061(97)00018-9}{10.1016/S0016-7061(97)00018-9}
  
Odgers, N.P., Sun, W., McBratney, A.B., Minasny, B., Clifford,
D., 2014. Disaggregating and harmonising soil map units through resampled
classification trees. Geoderma 214, 91--100. doi:
\href{https://doi.org/10.1016/j.geoderma.2013.09.024}{10.1016/j.geoderma.2013.09.024}
}
\seealso{
Other uncertainty functions: 
\code{\link{shannon_entropy}()}
}
\concept{uncertainty functions}
