% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsmart.R
\name{dsmart}
\alias{dsmart}
\title{Disaggregating and harmonising soil map units through resampled classification trees}
\usage{
dsmart(covariates, polygons, composition, n = 15, reals = 100, cpus = 1,
  obsdat = NULL)
}
\arguments{
\item{covariates}{A \code{RasterStack} of \emph{scorpan} environmental
covariates to both guide the C5 model fitting and provide the
necessary environmental data for the spatial mapping of predicted soil
classes.}

\item{polygons}{See \code{data(dsT_polygons)} for an example of 
requirements. Each polygon needs to have the following necessary attribute
data: A numeric unique map unit identifier, a character map unit name or 
code. And coupled columns of map unit classes and their respective 
proportions.}

\item{composition}{See \code{data(dsT_composition)} for an example of
requirements. First column indicates a numeric unique map unit identifier.
Second column indicates character map unit name or code. Third column 
indicates map unit compositions. And fourth column indicates the percentage
proportion of map unit compositions. All map unit codes and their subsequent
compositions are row stacked together.}

\item{n}{numeric; number of samples to take from each soil mapping polygon 
for C5 model fitting.}

\item{reals}{numeric; number of C5 modeling fitting and mapping realisations
to implement.}

\item{cpus}{numeric; number of compute nodes to use. Default is 1.}
}
\description{
This function, together with companion function \code{dsmartR} implements
the DSMART (Disaggregating and harmonising soil map units through resampled
classification trees) algorithm as described in Odgers et al. (2014). This
is the workhorse function that involves multiple resampling, C5 decision
tree model fitting, and subsequent mapping in order to realize potential
candidate soil classes within aggregated soil mapping units. There is also 
added facility to incorporate point observed data into the algorithm too.
}
