% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{order_stack_values}
\alias{order_stack_values}
\title{Order raster stack values}
\usage{
order_stack_values(r, cpus, n = nlayers(r))
}
\arguments{
\item{r}{A \code{RasterStack} whose layers contain the values to be ordered.}

\item{n_prob}{}
}
\value{
A \code{RasterStack} containing the sorted data.
}
\description{
This function orders the values in a raster stack. Ordering is performed on
the vector of values at each grid cell. See \code{\link[base]{order}} for
more information about how ordering works.
}
\details{
The layers in the returned raster stack correspond to the rank ordering of
the values in the original raster stack \code{r}. For example, the values in
the first returned layer identify the layer index of the original raster
stack in which the largest value was found; the values in the second returned
layer identify the original layer index in which the second largest value was
found, and so-on.
}
