% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.r
\name{sort_stack_values}
\alias{sort_stack_values}
\title{Sort raster stack values}
\usage{
sort_stack_values(r, cpus, n = nlayers(r), decreasing = TRUE)
}
\arguments{
\item{r}{A \code{RasterStack} whose layers contain the values to be sorted.}

\item{cpus}{An integer that identifies the number of CPU nodes for parallel
processing.}

\item{n}{An integer that identifies the number of layers in \code{r} that
should be sorted. Default is to sort all layers in \code{r}.}

\item{decreasing}{A boolean that indicates whether values should be sorted in
decreasing order (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
A \code{RasterStack} containing the sorted data.
}
\description{
#' This function sorts the values in a raster stack. Sorting is performed on
the vector of values at each grid cell. See \code{\link[base]{sort}} for more
information about how ordering works.
}
